rename to index.server
